Clean up docstring and add descriptions to parameters (#1723)
If we have functions that users provide or 3rd parties with a lot of
arguments and detailed docstrings, the description we extract is WAY too
long and is rejected by OpenAI.

This ensures we follow the specification exactly.

**Note**: One possible side-effect here is an increase in token usage,
but we are just moving information we already did send to a different
place, so the effect should be very small.

Below is from OpenAIs docs:

<img width="715" alt="Screenshot 2025-01-08 at 14 17 35"
src="https://github.com/user-attachments/assets/74487392-b891-4eb9-9e68-7947db1f52e4"
/>

Also fixes #1729